Moved orderSummary into it's own file - increased modularity
